requirements.txt
asgiref==3.5.2
Django==3.2.18
django-tinymce==3.6.1
Pillow==9.2.0
psycopg2-binary==2.9.5
sqlparse==0.4.2
